requirements.txt
Flask==2.3.3
numpy==1.23.5
opencv-python==4.8.0.74
Pillow==9.5.0
tensorflow==2.13.0
